Fixed wildcard alphabetical order
